redis-consumer/requirements.txt
redis==5.0.3
psycopg2-binary==2.9.9
prometheus_client
